requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-limiter==3.12
pydantic==2.10.6
bcrypt==4.2.1
python-dotenv==1.0.1
gunicorn==23.0.0
yfinance==0.2.40
beautifulsoup4==4.12.3
requests==2.32.3
